# client-api/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Matrix Client-Server Api",
    "version" : "r0.3.0",
    "title" : "Matrix Client-Server Api",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Anatoly",
      "url" : "https://github.com/ma1uta/jeon",
      "email" : "sablintolya@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "matrix.org:8448",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/_matrix/client/r0/keys/query" : {
      "post" : {
        "tags" : [ "Encryption" ],
        "summary" : "Returns the current devices and identity keys for the given users.",
        "description" : "",
        "operationId" : "query",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "JSON body request",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/QueryRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The device information.",
            "schema" : {
              "$ref" : "#/definitions/QueryResponse"
            }
          }
        }
      }
    },
...
    "/_matrix/client/r0/rooms/{roomId}/context/{eventId}" : {
      "get" : {
        "tags" : [ "EventContext" ],
        "summary" : "This API returns a number of events that happened just before and after the specified event.",
        "description" : "This allows clients to get the context surrounding an event.",
        "operationId" : "context",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "description" : "The room to get events from.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "eventId",
          "in" : "path",
          "description" : "The event to get context around.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The maximum number of events to return. Default: 10.",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "The events and state surrounding the requested event.",
            "schema" : {
              "$ref" : "#/definitions/EventContextResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "DeviceKeys" : {
      "type" : "object",
      "required" : [ "algorithms", "device_id", "keys", "signatures", "user_id" ],
      "properties" : {
        "algorithms" : {
          "type" : "array",
          "description" : "The encryption algorithms supported by this device.",
          "items" : {
            "type" : "string"
          }
        },
        "keys" : {
          "type" : "object",
          "description" : "Public identity keys. The names of the properties should be in the format &lt;algorithm&gt;:&lt;device_id&gt;. The keys themselves should be encoded as specified by the key algorithm.",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "signatures" : {
          "type" : "object",
          "description" : "Signatures for the device key object. A map from user ID, to a map from (algorithm):(device_id) to the signature.",
          "additionalProperties" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "unsignedDeviceInfo" : {
          "description" : "Additional data added to the device key information by intermediate servers, and not covered by the signatures.",
          "$ref" : "#/definitions/UnsignedDeviceInfo"
        },
        "user_id" : {
          "type" : "string",
          "description" : "The ID of the user the device belongs to. Must match the user ID used when logging in."
        },
        "device_id" : {
          "type" : "string",
          "description" : "The ID of the device these keys belong to. Must match the device ID used when logging in."
        }
      },
      "description" : "Device keys."
    },
...
    "Event" : {
      "type" : "object",
      "required" : [ "event_id", "origin_server_ts", "room_id", "sender", "state_key", "type" ],
      "properties" : {
        "content" : {
          "description" : "The fields in this object will vary depending on the type of event. When interacting with the REST API, this is the HTTP body.",
          "$ref" : "#/definitions/EventContent"
        },
        "type" : {
          "type" : "string",
          "description" : "The type of event. This SHOULD be namespaced similar to Java package naming conventions."
        },
        "sender" : {
          "type" : "string",
          "description" : "Contains the fully-qualified ID of the user who sent this event."
        },
        "unsigned" : {
          "description" : "Contains optional extra information about the event.",
          "$ref" : "#/definitions/Unsigned"
        },
        "event_id" : {
          "type" : "string",
          "description" : "The globally unique event identifier."
        },
        "room_id" : {
          "type" : "string",
          "description" : "The ID of the room associated with this event."
        },
        "origin_server_ts" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Timestamp in milliseconds on originating homeserver when this event was sent."
        },
        "prev_content" : {
          "description" : "The previous content for this event. If there is no previous content, this key will be missing.",
          "$ref" : "#/definitions/EventContent"
        },
        "state_key" : {
          "type" : "string",
          "description" : " A unique key which defines the overwriting semantics for this piece of room state. This value is often a zero-length string. The presence of this key makes this event a State Event. The key MUST NOT start with '_'."
        },
        "invite_room_state" : {
          "type" : "array",
          "description" : "This contains an array of StrippedState Events. These events provide information on a subset of state events such as the room name.",
          "items" : {
            "$ref" : "#/definitions/StrippedState"
          }
        }
      },
      "description" : "Event."
    },
    "EventContent" : {
      "type" : "object",
      "discriminator" : "type",
      "description" : "Parent class of all event contents part in the Event."
    },
...
    "EventContextResponse" : {
      "type" : "object",
      "properties" : {
        "start" : {
          "type" : "string",
          "description" : "Pagination token for the start of the chunk."
        },
        "end" : {
          "type" : "string",
          "description" : "Pagination token for the end of the chunk."
        },
        "profile_info" : {
          "type" : "object",
          "description" : "The historic profile information of the users that sent the events returned.",
          "additionalProperties" : {
            "$ref" : "#/definitions/Profile"
          }
        },
        "events_before" : {
          "type" : "array",
          "description" : "Events just before the result.",
          "items" : {
            "$ref" : "#/definitions/Event"
          }
        },
        "events_after" : {
          "type" : "array",
          "description" : "Events just after the result.",
          "items" : {
            "$ref" : "#/definitions/Event"
          }
        }
      },
      "description" : "Event context."
    },
...
    "Profile" : {
      "type" : "object",
      "properties" : {
        "displayname" : {
          "type" : "string",
          "description" : "The user's display name if they have set one, otherwise not present."
        },
        "avatar_url" : {
          "type" : "string",
          "description" : "The user's avatar URL if they have set one, otherwise not present."
        }
      },
      "description" : "JSON body for profile api (profile)."
    },
...
    "QueryRequest" : {
      "type" : "object",
      "required" : [ "device_keys" ],
      "properties" : {
        "timeout" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "The time (in milliseconds) to wait when downloading keys from remote servers. 10 seconds is the recommended default."
        },
        "token" : {
          "type" : "string",
          "description" : "If the client is fetching keys as a result of a device update received in a sync request, this should be the 'since' token of that sync request, or any later sync token. This allows the server to ensure its response contains the keys advertised by the notification in that sync."
        },
        "device_keys" : {
          "type" : "object",
          "description" : "The keys to be downloaded. A map from user ID, to a list of device IDs, or to an empty list to indicate all devices for the corresponding user.",
          "additionalProperties" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "description" : "JSON body parameters for Returns the current devices and identity keys for the given users."
    },
    "QueryResponse" : {
      "type" : "object",
      "properties" : {
        "failures" : {
          "type" : "object",
          "description" : "If any remote homeservers could not be reached, they are recorded here. The names of the properties are the names of the unreachable servers.",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "device_keys" : {
          "type" : "object",
          "description" : "Information on the queried devices. A map from user ID, to a map from device ID to device information. For each device, the information returned will be the same as uploaded via /keys/upload, with the addition of an unsigned property.",
          "additionalProperties" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/definitions/DeviceKeys"
            }
          }
        }
      },
      "description" : "Query response."
    },
...
    "StrippedState" : {
      "type" : "object",
      "required" : [ "content", "state_key", "type" ],
      "properties" : {
        "content" : {
          "type" : "object",
          "description" : "The content for the event.",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "type" : {
          "type" : "string",
          "description" : "The type for the event."
        },
        "state_key" : {
          "type" : "string",
          "description" : "The state_key for the event."
        }
      },
      "description" : "Provides information on a subset of state events such as the room name."
    },
...
    "Unsigned" : {
      "type" : "object",
      "properties" : {
        "age" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "The time in milliseconds that has elapsed since the event was sent. This field is generated by the local homeserver, and may be incorrect if the local time on at least one of the two servers is out of sync, which can cause the age to either be negative or greater than it actually is."
        },
        "redacted_because" : {
          "description" : "The event that redacted this event, if any.",
          "$ref" : "#/definitions/Event"
        },
        "transaction_id" : {
          "type" : "string",
          "description" : "The client-supplied transaction ID, if the client being given the event is the same one which sent it."
        },
        "prev_content" : {
          "type" : "object",
          "description" : "The previous content for this state. This will be present only for state events appearing in the timeline. If this is not a state event, or there is no previous content, this key will be missing.",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      },
      "description" : "Unsigned addition data."
    },
    "UnsignedDeviceInfo" : {
      "type" : "object",
      "properties" : {
        "device_display_name" : {
          "type" : "string",
          "description" : "The display name which the user set on the device."
        }
      },
      "description" : "Unsigned device info."
    },
...
  }
}